+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/client/app/components/WaspApp.tsx
@@ -1,13 +1,11 @@
-import { use, type ReactNode } from 'react'
+import { type ReactNode } from 'react'
 import { QueryClientProvider } from '@tanstack/react-query'
 
-import { queryClientInitialized } from '../../operations/index'
+import { queryClient } from '../client-setup'
 
 import { WebSocketProvider } from '../../webSocket/WebSocketProvider'
 
 export function WaspApp({ children }: { children: ReactNode }) {
-  const queryClient = use(queryClientInitialized)
-
   return (
     <QueryClientProvider client={queryClient}>
       <WebSocketProvider>