+++ waspc/e2e-tests/test-outputs/snapshots/wasp-build-golden/wasp-app/.wasp/out/sdk/wasp/client/app/client-setup.ts
@@ -1,0 +1,7 @@
+import { initializeQueryClient } from '../operations/queryClient'
+
+
+// PRIVATE API (framework code)
+// The QueryClient must be created only after the setup function completes,
+// because the setup function may call `configureQueryClient`.
+export const queryClient = initializeQueryClient()