+++ waspc/e2e-tests/test-outputs/snapshots/wasp-build-golden/wasp-app/.wasp/out/sdk/wasp/client/operations/index.ts
@@ -17,8 +17,6 @@ export {
     configureQueryClient,
     // PRIVATE API (framework code)
     initializeQueryClient,
-    // PRIVATE API (framework code)
-    queryClientInitialized
 } from './queryClient'
 
 export {